+++ assembly/deserialize/naive/unsigned.ts
@@ -14,8 +14,9 @@ export function deserializeUnsignedField_NAIVE<T extends number>(
   dstOffset: usize = 0,
 ): usize {
   const fieldPtr = dstObj + dstOffset;
+  if (srcStart >= srcEnd) return 0;
   let digit = <u32>load<u16>(srcStart) - 48;
-  if (digit > 9) unreachable();
+  if (digit > 9) return 0;
 
   if (sizeof<T>() == sizeof<u8>()) {
     let value: u64 = digit;