+++ assembly/deserialize/swar/float.ts
@@ -221,7 +221,10 @@ export function deserializeFloatField_SWAR<T extends number>(
   const fracDigits = <i32>((p - fracStart) >> 1);
 
   const mantDigits = intDigits + fracDigits;
-  if (mantDigits == 0) unreachable();
+  if (mantDigits == 0) {
+    if (JSON_STRICT) return 0;
+    unreachable();
+  }
 
   let exponent: i32 = -fracDigits;
 
@@ -254,6 +257,10 @@ export function deserializeFloatField_SWAR<T extends number>(
         exp = exp * 10 + <i32>d;
         expDigits++;
         if (expDigits > 4) {
+          do {
+            p += 2;
+            if (p >= srcEnd) break;
+          } while (<u32>load<u16>(p) - ASCII_ZERO <= 9);
           fallbackField<T>(origStart, p, fieldPtr);
           return p;
         }